moved issues-option to global folder
